+++ qa-include/ajax/wallpost.php
@@ -41,7 +41,7 @@ if ($errorhtml || !strlen($message) || !qa_check_form_security_code('wall-' . $t
 		$touseraccount['userid'], $touseraccount['handle'], $message, '');
 	$touseraccount['wallposts']++; // won't have been updated
 
-	$usermessages = qa_db_select_with_pending(qa_db_recent_messages_selectspec(null, null, $touseraccount['userid'], true, qa_opt('page_size_wall')));
+	$usermessages = qa_db_select_with_pending(qa_db_recent_messages_selectspec(null, null, $touseraccount['userid'], true, (int)qa_opt('page_size_wall')));
 	$usermessages = qa_wall_posts_add_rules($usermessages, 0);
 
 	$themeclass = qa_load_theme_class(qa_get_site_theme(), 'wall', null, null);
